app/requirements.txt
transformers
accelerate
invisible-watermark>=0.2.0
diffusers>=0.19.0
streamlit
pillow

